backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-dotenv==1.0.0
alembic==1.13.0
slowapi==0.1.9
redis==5.0.1
python-multipart==0.0.6
